Moved row.dart to lib/src.
Fixes #52
